backend/requirements.txt
django>=5.0,<7.0
djangorestframework>=3.15
django-cors-headers>=4.0
drf-spectacular>=0.27
posthog>=3.0
python-dotenv>=1.0
channels>=4.0
daphne>=4.0
